Fix Float/Double handling
Metadata Keys which specifically supported Float & Double types were not showing up in.
This was the case in 5.3. I didn't test in other engine versions.
This was because a Floats and Doubles are actually Reals, with the subtype of Float/Double. These keys were failing to be supported because `Key.DoesSupportProperty(Property)` would return false. Because it requires the type and subtype to match.

The config was updated to initialize these as `{ UEdGraphSchema_K2::PC_Real, UEdGraphSchema_K2::PC_Float }` and `{ UEdGraphSchema_K2::PC_Real, UEdGraphSchema_K2::PC_Double }`

The constructor for the struct was updated to `FixUp()` the values if they are being initialized incorrectly by code.

The config was updated to also call `FixUp()` on `PostInitProperties()` so that configurations which were saved with previous versions would be fixed also.